Fixed crash related to Log when out of memory
When the copter got out of memory and a log-block was created it hanged. This
patch fixes it by checking if the timer object created for each log block has
been successfully created.

Also added a mutex that ensure that a log block is not modified/deleted
while it is used.