Initial commit for the proximity measurement subsystem. The intention is for this subsystem to provide a common API for proximity sensors. The proximity subsystem runs in a separate task, currently running at 10Hz but this is configurable.
Initially the MaxBotix Sonar Range Finder is supported, and the driver is included in mb.[hc].

The proximity subsystem reports distance measurements as well as accuracies of measurements (as reported by drivers).
The current limitation is that only a single proximity sensor is supported at any given time. Currently, that is the
MaxBotix driver.

The proximity system includes a simple function for calculating the average of samples in a sliding window. The sliding
window is configurable for number of samples in window.

Distance measurements are made available to cfclient by both the MaxBotix driver as well as by the proximity subsystem.
The latter also makes the average distance measurement available to the cfclient.

This commit disables the proximity system by default (at compile time). Enable by uncommenting the following defines
in config.h: PROXIMITY_ENABLE and MB_ENABLE