CRTP: defensive asserts in crtpSend*.
These asserts are going to be performed in radiolink, but at that point
the packet has passed through a queue, making it somewhat difficult to
identify the source of the corrupt packet.

Here I've replicated the asserts at the crtpSend* functions, which are
the primary ways that packets enter that queue.  It is now easy to
identify the origin of corrupt packets by setting a breakpoint at
'assertFail'.

This does *not* provide fault isolation: an assert in a low-priority
task may still starve higher-priority tasks because of contention over
shared resources.  This is a larger problem that stems from the
firmware's current use of timeouts (portMAX_DELAY) and the
implementation of the worker task.

But this is a start.