log: don't send huge packets.
Before this commit, the log subsystem would blindly follow the
instructions it received from the client, even if it meant generating a
packet that was over CRTP_MAX_DATA_SIZE in length.  The errant packet
would wind up killing off either the radiolink task (before the previous
commit) or the worker task (as of the previous commit).

Either way, it kills telemetry and turns on the failure lights.  Control
authority (one-way) is retained.

As of this commit, log packets will be truncated at the last item that
fits completely into the buffer.  This confuses the Python client's
plotter implementation, but leaves flight telemetry and control
authority intact.  As soon as the log configuration is changed to remove
the excess data, service is restored.

Aside:

This is a classic buffer overrun, and because the packet is
stack-allocated, it's also a stack smash.

The case where the packet was allowed to escape into radiolink is
particularly interesting, because it implies that logRunBlock is
permitted to return using the smashed stack into a corrupt worker task
loop.  This is likely exploitable.