Added ARM's CMSIS-DSP lib to the CF2 build
The lib is added with an other Makefile that generates a .a file. This is to be
able to compile the DSP lib with -O2 optimisation flag even when the
firmware is compiled in debug mode (-O0 optimisation).

make clean is not cleaning the lib. To clean all the lib files run
make mrproper

Also changed the CMSIS header file to the new version. Since we only use
these headers form CMSIS it does not seems to cause any problems.

Closes #129