Added quick and dirty position set mode to the commander
This is waiting for a proper implementation of #113. The position setpoint
is mapped to:
x: Roll
y: -1*Pitch
z: Thrust/1000
Yaw: Yaw!

The mode is activated by the paramter flightmode.posSetMode. A Z
setpoint of 0 turns off the motors and since the Z setpoint is mapped on the
uint16 thrust the maximal altitude is limited to about 65m.