Added crtp localization service.
The current external positining service has been renamed to localization
service to handle a brauder range of requests/data. A new channel 1 has
been allocated to handle generic data where the firs byte defines this.
The first type implemented is the RANGE_STRAM_FLOAT which sends a stream
of ranges in floats together with the achor ID. The stream is enabled by
setting the parameter locSrv.enRangeStreamFP32 to 1 and when a Loco Deck
is connected.

Closes #196