Cleanup and final implementation of proptest code. Closes #370
A initial implementation has been made and it works like this.
- To run the propeller test the parameter health.runPropTest should be
set to 1. This can be done using the cfclient or using the python lib
parameter framework e.g. cf.param.set_value('health.runPropTest', '1').
Setting it again will run another new test.
- The test will then run by spinning each propeller after another to
check the vibration. If a propeller is found to be over the limit (set
in config.h as PROPELLER_BALANCE_TEST_THRESHOLD) the failed propeller
will make three short beeps and a sound so it can be identified.
- In the console in the cfclient information of the test will be printed
and also if any propeller failed.
- After a test has run it might be wise to restart the CF2.