#375 Improved outlier filter for TDoA
The call to the outlier filter has been moved to the kalman filter.
The outlier filter is based on the estimated distance from the current estimated position to the candidate sample.
A number of counters are used to keep track of how many samples that are within certain error levels.
The counters in turn are used to determine which error level to accept and if a sample should be passed on to the kalman filter or not.