uSD card: Add synchronous logging mode.
The synchronous logging mode can log variables synchronously in the
stabilizer loop. This helps to analyze flight crashes etc. where
the order and consistency between variables is crucial. The older
(asynchronous mode) is retained, although the config file format
changed slightly.

Tested on a CF 2.0 (both synchronous and asynchronous modes using
the config.txt that is part of this repository.)