Add logging variables for compressed state and setpoint.
This will allow to create a log block that can capture position,
orientation (and if desired vel/acc) in a single, time-syncronized
block. That is important for accurate pose reporting (e.g., for
roadrunner, or for CFs that are not controlled with external
mocap).

The compressed setpoint allows to capture both estimated and
desired pose, which is important to compute metrics such as
tracking errors.