#423 Added DMA Tx functions to UART1 & UART2.
The DMA functions will free up CPU time as it will not wait in a
spin-loop. The DMA stream for UART1 conflicts with CF-RZR SPI2 so this
can not be used together, thus the UART1 DMA functionality has to be
enabled with ENABLE_UART1_DMA. The functionality has been tested with
uart1test.c and uart2test.c.