#461: Keep track of number of syncrhornized BS
If synchronized on two basestations, do not allow to synchronize on one
basestation. If a second basestation was never seen, the
synchronization will work for one basestation.

The aim is to allow both running from a single LHv1 basestation
and protect two basestation systems from wrongly using the slave
basestation as master after desynchronization