Fixed fatfs_sd integration layer and more. Closes #270
Finally the SD-card can be used together with Flow and LPS deck. This
issue fixes the fatfs_sd integration layer that was flawed. The SPI
mutex was not taken and released in a proper manner.

The sd-card wite task is running at a very low priority so when the
system is heavily loaded the writing will not keep up. This often
happens if LPS is in TWR mode and logging is on because TWR uses the SPI
bus extensively, almost 15k accesses per second.