#700 Check lighthouse bitstream using CRC
Removes the lighthouse deck bitstream from the fimware and check if the
bitstream in the deck is supported using CRC instread of a direct
byte comparison. This allows to remove the DISABLE_LIGHTHOUSE_DRIVER
macro and so to have the listhouse always enabled.

This firmware is now incapable of flashing a new bitstream if needed.
Implementation of a flashing over CRTP is needed to make it fonctional
again.