Copy functional uSD deck documentation from wiki to repo documentation.
This 1:1 copy will ensure that the history of the documentation is kept.
The goal is to delete the functional part of the text in the wiki and
refer to the repo instead. Future functional changes, e.g., the new
event-trigger-based logging, can than be documented directly with the
firmware changes.