Add supervisor module
This module will keep track of Crazyflie state. The first two state
variables added are:

  function              description         log variable
  ----------------------------------------------------------
  supervisorCanFly    |  Will return true  | sys.canfly
                      |  if the Crazyflie  |
                      |  is able to fly    |
  ----------------------------------------------------------
  supervisorIsFlying  |  Will return true  | sys.isFlying
                      |  if we belive that |
                      |  the Crazyflie is  |
                      |  currently flying  |
  ----------------------------------------------------------

We attempt to reduce the places where the firmware tries to figure out
if the Crazyflie is flying or not. The plan is that this is the job for
the supervisor module. Right now the implementation is naive and only
looks at wether the thrust of the motors are above a certain threshold.
The implementation is taken from the sitaw module. There is certainly
room for improvement.

The CanFly state is the same as before, but lifted into the supervisor
module. We say that the Crazyflie can fly if it is not connected to the
charger. There may be other criterion we can add later on.