Lighthouse: Allow to set system type at runtime
System type (Lighthouse V1 or V2) can be set with a parameter. The new
setting is stored in persistent storage and loaded on next statups.
When changing types, the current configuration (geometry and
calibration) is cleared from RAM.

By default lighthouse V2 is used if no setting exists in storage.