Lighthouse crossing beam: improve result quality (#733)
The previous code reported the position of a single sensor in
lighthouse.{x,y,z}, rather than the position w.r.t. the center of mass
of the Crazyflie. This new code now averages over all sensors.

The lighthouse.delta variable gives an indication of the quality of the
result. This indication does not correlate with absolute accuracy (when
measured using a motion capture system as ground truth). However, it
provides a helpful indication on whether lighthouse.{x,y,z} is valid
(deltaLog>0) or not (deltaLog == 0).

Co-authored-by: Wolfgang Hoenig <whoenig@caltech.edu>