Lighthouse2: Filter stale measurements that occur during interference
Since LH2 basestations are not synchronized per cable, there is a
natural interference when beams reach the sensors at the same time.
This causes stale measurements (re-using of old angles), which result
in incorrect position estimation during motion.

This change invalidates old sweeps based on the known frequency of the
LH2 laser sweeps.