Add graceful shutdown support
Currently to shut-down the Crazyflie, the nRF51 simply cut the power
to thae system. This was working fine until we started implementing
SD-card functionality: when shutting down the system that way we have
huge chances of corrupting the SD-card filesystem. This has lead to
implementation of very inefficient sd-car logging procedure to decrease
the probability of corruption.

A better solution for the shutdown problem would be to implement a
graceful shutdown in the system ...

  ... when the nRF51 wants to shutdown the system (ie. after a button
      press), it sends a message to the stm32 requesting shutdown

  ... in the stm32, modules that wants to be notified of shutdown are
      called in sequence and return when they are ready for shutdown

  ... when all the module have returned, the stm32 sends to the nRF51 a
      "ready for shutdown" message

  ... the nRF51 cuts the power. If no "ready to shutdown" was received
      after a timeout, the power is cut anyway

Closes #623