Refactor controller update to take const setpoint
The update function of controllers should not alter the setpoint itself.
This could lead to unexpected behavior and makes the Python interface for
software-in-the-loop simulation less clean.

The change itself is simple, apart from the PID velocity controller.
Here, there is a small behavior change: when logging, the user-specified setpoint will be shown, not
the intermedate value altered by the controller.
However, this seems to be the more logical behavior and is consistent
with the behavior of all other controllers that compute intermediate
setpoints (e.g., position controller of mellinger
 computes intermediate angles).