Storage; Defragment kve storage when starting up by default
The kve will be defragmented before running the kve sanity check.
Defragmenting before checking will make sure that the kve is sane after defragmentation
and if not the kve will be reformatted.

This can be turned of in the Kconfig, so that users that do not want this to possibly
change their startup time (if eeprom is severely fragmented) can go back to only defragmenting
once kve is full