Modify rpm deck driver test to work properly on Brushless motors
The ESCs needs to be fed a signal constantly to not go to sleep,
however we can not start the test until ESCs are started (which we dont
know exactly when that is).

To get around this issue we will wait a time, while constantly feeding the
ESCs at 1KHz with a 0 PWM signal. After this time we start the test.