Force manual arming for brushless
This automatically forces manual arming if users manually select the brushless as the platform in kbuild instead of by loading the brushless defconfig.
Without this, users could compile a flyable firmware and accidentally forget to enable manual arming.
People that want autoarming on the brushless platform will have to modify the kbuild system.